+++ middleware/client.js
@@ -8,6 +8,7 @@ const authTypes = require('../config/auth').types;
 const privilegedRoles =  require('../config/roles').privilegedRoles;
 const authTypesConfig = require('../config').authTypes
 const defaultRole =  require('../config/roles').defaultRole;
+const getClientIdFromRequest = require('../utils/getClientIdFromRequest');
 
 exports.withAll = (req, res, next) => {
   Client
@@ -21,7 +22,7 @@ exports.withAll = (req, res, next) => {
 }
 
 exports.withOne = (req, res, next) => {
-  let clientId = req.body && req.body.clientId ? req.body.clientId : req.query.clientId;
+  const clientId = getClientIdFromRequest(req);
   
   if (!clientId) {
     clientId = req.query.client_id;